requirements.txt
# --- Core Deep Learning Libraries ---
torch==2.3.0
torchvision==0.18.0
timm==0.9.12
tensorflow==2.17.0

# --- Model Conversion and Inference ---
onnx==1.16.0
onnxruntime==1.17.3

# --- Image Processing & Utilities ---
opencv-python==4.10.0.84
albumentations==1.4.4
Pillow==10.3.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.0
scikit-learn==1.5.0
tqdm==4.66.4

# --- Web Framework (Flask App) ---
flask==3.0.3
flask-cors==4.0.1
requests==2.32.3

# --- Optional: Development Tools ---
jupyterlab==4.2.0
